Move deposit and vote logic to types package
