+++ x/gov/types/deposit.go
@@ -1,0 +1,43 @@
+package types
+
+import (
+	"fmt"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
+
+// Deposit
+type Deposit struct {
+	Depositor  sdk.AccAddress `json:"depositor"`   //  Address of the depositor
+	ProposalID uint64         `json:"proposal_id"` //  proposalID of the proposal
+	Amount     sdk.Coins      `json:"amount"`      //  Deposit amount
+}
+
+func (d Deposit) String() string {
+	return fmt.Sprintf("deposit by %s on Proposal %d is for the amount %s",
+		d.Depositor, d.ProposalID, d.Amount)
+}
+
+// Deposits is a collection of Deposit objects
+type Deposits []Deposit
+
+func (d Deposits) String() string {
+	if len(d) == 0 {
+		return "[]"
+	}
+	out := fmt.Sprintf("Deposits for Proposal %d:", d[0].ProposalID)
+	for _, dep := range d {
+		out += fmt.Sprintf("\n  %s: %s", dep.Depositor, dep.Amount)
+	}
+	return out
+}
+
+// Equals returns whether two deposits are equal.
+func (d Deposit) Equals(comp Deposit) bool {
+	return d.Depositor.Equals(comp.Depositor) && d.ProposalID == comp.ProposalID && d.Amount.IsEqual(comp.Amount)
+}
+
+// Empty returns whether a deposit is empty.
+func (d Deposit) Empty() bool {
+	return d.Equals(Deposit{})
+}
